requirements.txt
fastapi
uvicorn[standard]
gunicorn
pymongo
python-dotenv


